pyproject.toml
[tool.poetry]
name = "crewai-rag-basics"
version = "0.1.0"
description = ""
authors = ["bhancockio <brandon@brandonhancock.io>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "0.36.0"}
python-dotenv = "1.0.0"
crewai-tools = "^0.4.7"
firecrawl-py = "^0.0.20"
embedchain = {extras = ["youtube"], version = "^0.1.116"}

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
